Moved tests files to the root
