+++ packages/cli/src/publish.test.ts
@@ -1,12 +1,12 @@
 import { CannonWrapperGenericProvider, copyPackage, DeploymentInfo, IPFSLoader, OnChainRegistry } from '@usecannon/builder';
 import * as builder from '@usecannon/builder';
-import { LocalLoader } from '../loader';
-import { publish } from './publish';
+import { LocalLoader } from './loader';
+import { publish } from './commands/publish';
 import { ethers } from 'ethers';
 import fs from 'fs-extra';
 import path from 'path';
-import { resolveCliSettings } from '../settings';
-import * as settings from '../settings';
+import { resolveCliSettings } from './settings';
+import * as settings from './settings';
 import _ from 'lodash';
 
 describe('publish command', () => {
